resistances should be short, not byte and rare needs to be sbyte

+++ OpenNos.DAL.EF.MySQL/Entities/Item.cs
@@ -58,7 +58,7 @@ namespace OpenNos.DAL.EF.MySQL
 
         public byte DarkElement { get; set; }
 
-        public byte DarkResistance { get; set; }
+        public short DarkResistance { get; set; }
 
         public short DefenceDodge { get; set; }
 
@@ -82,7 +82,7 @@ namespace OpenNos.DAL.EF.MySQL
 
         public byte FireElement { get; set; }
 
-        public byte FireResistance { get; set; }
+        public short FireResistance { get; set; }
 
         public short HitRate { get; set; }
 
@@ -120,7 +120,7 @@ namespace OpenNos.DAL.EF.MySQL
 
         public byte LightElement { get; set; }
 
-        public byte LightResistance { get; set; }
+        public short LightResistance { get; set; }
 
         public short MagicDefence { get; set; }
 
@@ -183,7 +183,7 @@ namespace OpenNos.DAL.EF.MySQL
 
         public byte WaterElement { get; set; }
 
-        public byte WaterResistance { get; set; }
+        public short WaterResistance { get; set; }
 
         #endregion
     }
